Move main buffers to macros
